#6 - Springboot 프로젝트 생성했을 때와 비슷한 구조
